+++ src/adapter/http/src/data/mod.rs
@@ -8,6 +8,7 @@
 // by the Apache License, Version 2.0.
 
 mod ingest_handler;
+pub mod metadata_handler;
 mod query_handler;
 pub mod query_types;
 mod router;